go-tour: tour redesign
The web UI has been implemented from scratch using AngularJS and fetches the content of the modules via a REST APi.

The tour has been split into different modules composed by a list of lessons.

When a new lesson is added it will be accessible publicly, but it won't appear in the list of modules until the corresponding `lesson` element is added to list.html.

The new version is accessible at http://campoy.cat/gotour

LGTM=adg
R=r, adg, kamil.kisiel, egonelbre, nightlyone, dan.kortschak, gobot, bradfitz, davidday
CC=golang-codereviews
https://golang.org/cl/36500046