content: add an extra slide for slices of slice
The `Pic` exercise throws people a bit, since they've not seen
the syntax for slices of slice.

This shows syntax that is similar enough to provide a hint, without
giving away the answer. They still have to figure out the assignment in
the loop on their own.

Fixes golang/go#12284

Change-Id: I9f12e97c4ca0ac37d08898f8a898c78b2f983ef4
Reviewed-on: https://go-review.googlesource.com/16028
Reviewed-by: Andrew Gerrand <adg@golang.org>