content: rename variable names stringer exercise
Presumably n was for name, but it wasn't immediately clear
since "name" was not mentioned anywhere in the code example.

There was a bit of discussion about whether it should be called
k (for key) or s (for string) or h (for host).

It seems like a lot of the confusion stems from the map being named
addrs, while actually containing both the name and the ip address of
the hosts.

Change-Id: Idf4483f142831b375cb0b1a5ebd58e8f3d9df30b
Reviewed-on: https://go-review.googlesource.com/19405
Reviewed-by: Andrew Gerrand <adg@golang.org>