content: make welcome/3 slide more beginner friendly
Re-add the link to installing Go on the welcome/3 slide,
because it's a pre-requisite for being able to run go get.

This was included before, but got lost in a refactor after
the tour binary moved out of the Go distribution.

Additionally, make it more clear that installing the tour
locally is an optional step that only needs to be done if
one wishes to run the tour offline; it's not mandatory to
get through the tour. Running the tour online has a lower
overhead, and should help keep the tour accessible to more
people.

Fixes golang/go#34585
Updates golang/go#24819

Change-Id: If5f51e9e876722a86afca054bf78428f78bb3076
Reviewed-on: https://go-review.googlesource.com/c/tour/+/197857
Reviewed-by: Emmanuel Odeke <emm.odeke@gmail.com>