tree: document sorted nature of random trees
For the use that the tour makes of "golang.org/x/tour/tree" it is
important that the generated trees are sorted. While New() does generate
sorted trees it is current documented as "returns a new, random binary
tree" and many readers will assume "random" implies "non-sorted".

Update doc that trees are "randomly-structured, sorted" to clarify what
is random and that trees are sorted.