Update Welcome Section
Updated the Welcome Section of the Tour to mostly remove bits that we don't
want to worry about for our DDM Go Training. Like the Submit Feedback
button and the link to the the table of contents. The table of contents
is kinda confusing, and we might change it later, so it's easier to just
remove the references to it.